restructuring folder
Used the GEE package model to restrucutre .py files and the outer __init__.py file in the sandpyper dir.
