--- sandpyper/dynamics/__init__.py
@@ -1,9 +1,0 @@
-from .dynamics import (
-    attach_trs_geometry,
-    infer_weights,
-    get_coastal_Markov,
-    BCDs_compute,
-    steady_state_transect,
-    compute_rBCD_transects
-)
-__version__ = '0.0.1'
