--- sandpyper/hotspot/__init__.py
@@ -1,4 +1,0 @@
-from .hotspot import (
-    LISA_site_level
-)
-__version__ = '0.0.1'
