--- sandpyper/labels/__init__.py
@@ -1,7 +1,0 @@
-from .labels import (
-    get_sil_location,
-    plot_sil,
-    get_opt_k,
-    kmeans_sa
-)
-__version__ = '0.0.1'
