--- sandpyper/outils/__init__.py
@@ -1,23 +1,0 @@
-__version__ = "0.0.1"
-from .outils import (
-    # nmad, needs ASTROPY
-    polygonize_valid,
-    matchup_folders,
-    find_skiprows,
-    open_gcp_file,
-    timeseries_to_gdf,
-    gdf_distance_matrix,
-    getCrs_from_raster_path,
-    getCrs_from_transect,
-    cross_ref,
-    getListOfFiles,
-    extract_loc_date,
-    getListOfDate,
-    getDate,
-    getLoc,
-    getListOfFiles,
-    create_spatial_id,
-    create_id,
-    coords_to_points,
-    round_special,
-    filter_filename_list)
