--- sandpyper/profile/__init__.py
@@ -1,10 +1,0 @@
-__version__ = '0.0.1'
-
-from .profile import (
-    get_terrain_info,
-    get_elevation,
-    get_profiles,
-    get_dn,
-    get_profile_dn,
-    get_merged_table,
-    extract_from_folder)
