--- sandpyper/volumetrics/__init__.py
@@ -1,13 +1,0 @@
-from .volumetrics import (
-    getVol,
-    prep_heatmap,
-    fill_gaps,
-    interpol_integrate,
-    get_beachface_length,
-    get_m3_m_location,
-    new_get_state_vol_table,
-    new_get_transects_vol_table,
-    new_plot_alongshore_change,
-    plot_mec_evolution
-)
-__version__ = '0.0.1'
